Client/app/requirements.txt
numpy==1.24.3
kafka-python>=2.0.2,<3.0.0